\name{cases.suf.dcn}
\alias{cases.suf.dcn}
\title{
	List cases deviant with regards to consistency.
}
\description{
	Function extracts cases deviant with regards to consistency from an object of class "qca".
}
\usage{
cases.suf.dcn(results, outcome, solution = 1)
}
\arguments{
	\item{results}{
		An object of class "qca".
	}
	\item{outcome}{
		A character string with the name of the outcome.
	}
	\item{solution}{
		A numeric vector where the first number indicates the number of the
		solution according to the order in the "qca" object.
	}
}
\details{
}
\value{
}
\references{
}
\author{
	Juraj Medzihorsky
}
\note{
}
\seealso{
	\code{\link[QCA:eqmcc]{eqmcc}}
}
\examples{
\keyword{ QCA }
