\name{CUSFS}
\alias{CUSFS}
\docType{data}
\title{Thomann (2015)}
\description{The \code{CUSFS} data frame has 76 rows and 7 sets}
\usage{data(CUSFS)}
\format{
  A data frame with 76 observations on the following 7 sets.
  \describe{
    \item{\code{RESP}}{Cond.}
    \item{\code{SAL}}{Cond.}
    \item{\code{RES}}{Cond.}
    \item{\code{VPO}}{Cond.}
    \item{\code{VPL}}{Cond.}
    \item{\code{COERC}}{Cond.}
    \item{\code{CUSTOM}}{Outcome}
  }
}
\details{Data used by Thomann (2015) to explain customization. Data are fuzzy-sets.}

\references{
	
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012) How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at \url{www.cambridge.org/schneider-wagemann}.
}
\examples{data(CUSFS)}
\keyword{datasets}
