\name{FakeCS}
\alias{FakeCS}
\docType{data}
\title{Fake crisp-set data}
\description{The \code{FakeCS} data frame has 30 rows and 5 sets}
\usage{data(FakeCS)}
\format{
  A data frame with 30 observations on the following 5 sets.
  \describe{
    \item{\code{y}}{a numeric vector. Outcome with 2 categories (crisp-set).}
    \item{\code{j}}{a numeric vector. Condition with 2 categories (crisp-set).}
    \item{\code{z}}{a numeric vector. Condition with 2 categories (crisp-set).}
    \item{\code{w}}{a numeric vector. Condition with 2 categories (crisp-set).}
    \item{\code{k}}{a numeric vector. Condition with 2 categories (crisp-set).}
  }
}
\details{The data frame has only exercise purpuses to let the user learn how to perform crisp-set QCA in R.}
%\source{}
\references{Schneider, C. Q., Wagemann, C., Quaranta, M. (2012) How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at www.cambridge.org/schneider-wagemann.
	
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.}
\examples{data(FakeCS)}
\keyword{datasets}
