\name{JOBF}
\alias{JOBF}
\title{Fake fuzzy-set job motivation data.}
\description{Fake fuzzy-set job motivation data used for Chapter 3 of the Oana et al. (forthcoming) book.}
\usage{data(JOBF)}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{\code{HS}}{Condition: High salary}
    \item{\code{FS}}{Condition: Flexible schedule}
    \item{\code{LS}}{Condition: Lenient supervisor}
    \item{\code{LJS}}{Condition: Low job security}
    \item{\code{PP}}{Condition: High promotion potential}
    \item{\code{HM}}{Outcome: High motivation}
    \item{\code{FSorLS}}{Condition:}
  }
}
\references{
Oana, Ioana-Elena, Carsten Q. Schneider, and Eva Thomann (forthcoming). Qualitative Comparative Analysis (QCA) using R: A Gentle Introduction. Cambridge: Cambridge University Press.
}
\examples{data(JOBF)}
\keyword{datasets}
