\name{KA}
\alias{KA}
\docType{data}
\title{Koenig-Archibugi (2004)}
\description{The \code{KA} data frame has 13 rows and 5 variables}
\usage{data(KA)}
\format{
  A data frame with 13 observations on the following 5 variables.
  \describe{
    \item{\code{supranat}}{a numeric vector. Government support for supranational CFSP.}
    \item{\code{identmass}}{a numeric vector. European identity of the general public.}
    \item{\code{conform}}{a numeric vector. Policy conformity.}
    \item{\code{region}}{a numeric vector. Regional governance.}
    \item{\code{capab}}{a numeric vector. Material capabilities.}
  }
}
\details{Data are used by Koenig-Archibugi (2004) to analyze government preferences for istitutional change in EU foreign and security policy. Data are fuzzy-sets.}
%\source{}
\references{Koenig-Archibugi, M. (2004) "Explaining Government Preferences for Istitutional Change in EU Foreign and Security Policy", International Organization 58, pp.137-174.

Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012) How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at \url{www.cambridge.org/schneider-wagemann}}
\examples{data(KA)}
\keyword{datasets}
