\name{LIPR}
\alias{LIPR}
\docType{data}
\title{Lipset (1959), raw data}
\description{The \code{LIPR} data frame has 18 rows and 6 variables}

\usage{data(LIPR)}
\format{
  A data frame with 18 observations on the following 6 variables.
  \describe{
    \item{\code{SURVIVED}}{a numeric vector. Outcome, survival of democracy during the inter-war period.}
    \item{\code{DEVELOPED}}{a numeric vector. Condition, level of economic development.}
    \item{\code{URBAN}}{a numeric vector. Condition, level of urbanization.}
    \item{\code{LITERATE}}{a numeric vector. Condition, level of literacy.}
    \item{\code{INDUSTRIAL}}{a numeric vector. Condition, level of industrialization.}
    \item{\code{UNSTABLE}}{a numeric vector. Condition, politically stable countries.}
  }
}
\details{Data used by Ragin (2009) to illustrates the variants of QCA. Originally by Lipset (1959). Data are raw-scores.}
\references{
Lipset, S. M. (1959) "Some Social Requisites of Democracy: Economic Development and Political Legitimacy", American Political Science Review 53, pp. 69-105.

Ragin, C. C. (2009) "Qualitative Comparative Analysis. Using Fuzzy Sets (fsQCA)." In Rihoux, B., and Ragin, C. C. (eds.) Configurational Comparative Methods. Qualitative Comparative Analysis (QCA) and Related Techniques. Thousand Oaks, CA and London: Sage, pp. 87-121.

Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012) How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at \url{www.cambridge.org/schneider-wagemann}.}
\examples{data(LIPR)}
\keyword{datasets}
