\name{matches.suf.dcvirr}
\alias{matches.suf.dcvirr}
\title{
	Match cases deviant with regards to consistency with individually irrelevant cases.
}
\description{
	Matches cases deviant with regards to consistency with individually irrelevant cases.
}
\usage{
matches.suf.dcvirr(results, outcome, solution = 1, max_pairs = 5)
}
\arguments{
	\item{results}{
		An object of class "qca".
	}
	\item{outcome}{
		A character string with the name of the outcome.
	}
	\item{solution}{
		A numeric vector where the first number indicates the number of the
		solution according to the order in the "qca" object.
	}
	\item{max_pairs}{
		Maximum number of pairs to be extracted.
	}
}
\details{
}
\value{
}
\references{
}
\author{
	Juraj Medzihorsky
}
\note{
}
\seealso{
	\code{\link[QCA:eqmcc]{eqmcc}}
}
\examples{}
\keyword{QCA}
