\name{PAYF}
\alias{PAYF}
\docType{data}
\title{Paykani et al. (2018)}
\description{The \code{PAYF} data frame has 131 rows and 9 sets.  The data is calibrated into fuzzy sets.}
\usage{data(PAYF)}
\format{
  A data frame with 131 observations on the following 9 sets.
  \describe{
    \item{\code{COUNTRY}}{Country}
    \item{\code{REGION}}{Region the country belongs to.}
    \item{\code{HE}}{Condition: healthy education system}
    \item{\code{GG}}{Condition: good governance}
    \item{\code{AH}}{Condition: affluent health  system}
    \item{\code{HI}}{Condition: high income inequality}
    \item{\code{HW}}{Condition: high wealth}
    \item{\code{HL}}{Outcome: high life expectancy}
    \item{\code{LL}}{Negated Outcome: low life expectancy}

  }
}
\references{
Paykani, Toktam, Rafiey, Hassan, and Sajjadi, Homeira. 2018. A fuzzy set qualitative comparative analysis of 131 countries: which configuration of the structural conditions can explain health better? International journal for equity in health, 17(1), 10.
}
\examples{data(PAYF)}
\keyword{datasets}
