\name{PAYR}
\alias{PAYR}
\docType{data}
\title{Paykani et al. (2018)}
\description{The \code{PAYR} data frame has 131 rows and 9 variables. The data is raw, uncalibrated.}
\usage{data(PAYR)}
\format{
  A data frame with 131 observations on the following 9 sets.
  \describe{
    \item{\code{COUNTRY}}{Country}
    \item{\code{REGION}}{Region the country belongs to.}
    \item{\code{LIFEX}}{Life expectancy, raw data.}
    \item{\code{EDUC}}{Education, raw data.}
    \item{\code{GOV}}{Governance, raw  data.}
    \item{\code{HEAL}}{Health system, raw data.}
    \item{\code{INCEQ}}{Income inequality, raw data.}
    \item{\code{WEAL}}{Wealth, raw data.}
  }
}
\references{
Paykani, Toktam, Rafiey, Hassan, and Sajjadi, Homeira. 2018. A fuzzy set qualitative comparative analysis of 131 countries: which configuration of the structural conditions can explain health better? International journal for equity in health, 17(1), 10.
}
\examples{data(PAYR)}
\keyword{datasets}
