\name{Pennings}
\alias{Pennings}
\docType{data}
\title{Pennings (2003)}
\description{The \code{Pennings} data frame has 45 rows and 5 variables}
\usage{data(Pennings)}
\format{
  A data frame with 45 observations on the following 5 variables.
  \describe{
    \item{\code{K}}{a numeric vector. Outcome, constitutional control.}
    \item{\code{C}}{a numeric vector. Condition, consensus democracy.}
    \item{\code{P}}{a numeric vector. Condition, presidentialism.}
    \item{\code{N}}{a numeric vector. Condition, new democracy.}
    \item{\code{R}}{a numeric vector. Condition, rigid constritution.}
  }
}
\details{Data used by Pennings (2009) to explain constitutional control. Data are fuzzy-sets.}

\references{Pennings, P. (2009) "Beyond Dichotomous Explanations: Explaining Constitutional control of the Executive with Fuzzy-sets", European Journal of Political Research 42, pp. 541-567.
	
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012) How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at \url{www.cambridge.org/schneider-wagemann}.
}
\examples{data(Pennings)}
\keyword{datasets}
