\name{print.casestheoryeval}
\alias{print.casestheoryeval}
\title{Print a casestheoryeval Object}
\description{
  Information regarding the names and numbers of cases in the intersections between theory and empirics. 
}
\usage{
  \method{print}{casestheoryeval}(x, \dots)
}
\arguments{
  \item{x}{an object inheriting from class \code{cases.theory.evaluation}, representing
    the name and numbers of cases in the various intersections between theory and empirics.
  }
  \item{\dots}{optional arguments passed to \code{print.default}; see
    the documentation on that method function}
}
\references{
  Ragin, C. C. 1987. The Comparative Method: Moving Beyond Qualitative and Quantitative Strategies. Berkeley: University of California Press, pp. 118-121.
  
  Schneider, C. Q., Wagemann, C. 2012. Set-Theoretic Methods for the Social Sciences: A Guide to Qualitative Comparative Analysis. Cambridge: Cambridge University Press, chapter 11.3.
}
\author{
  Ioana-Elena Oana
}
\seealso{
  \code{\link[QCAGUI:eqmcc]{eqmcc}}
}
\keyword{QCAGUI}
