\name{print.matchessuf}
\alias{print.matchessuf}
\title{Print a matchessuf Object}
\description{
  Information regarding matching of cases following the INUS principle. 
}
\usage{
  \method{print}{matchessuf}(x, \dots)
}
\arguments{
  \item{x}{an object inheriting from class \code{matches.suf.typtyp} and \code{matches.suf.typiir}, representing
    matching of cases following the INUS principle.  
  }
  \item{\dots}{optional arguments passed to \code{print.default}; see
    the documentation on that method function}
}
\references{
  Ragin, C. C. 1987. The Comparative Method: Moving Beyond Qualitative and Quantitative Strategies. Berkeley: University of California Press, pp. 118-121.
  
  Schneider, C. Q., Wagemann, C. 2012. Set-Theoretic Methods for the Social Sciences: A Guide to Qualitative Comparative Analysis. Cambridge: Cambridge University Press, chapter 11.3.
}
\author{
  Ioana-Elena Oana
}
\seealso{
  \code{\link[QCA:minimize]{minimize}}
}
\keyword{QCA}
