\name{print.thintersect}
\alias{print.thintersect}
\title{Print a theory intersections object}
\description{
  Boolean intersections between theory and empirics. 
}
\usage{
  \method{print}{thintersect}(x, \dots)
}
\arguments{
  \item{x}{an object inheriting from class \code{theory.intersections}, representing
    the various intersections between theory and empirics.
  }
  \item{\dots}{optional arguments passed to \code{print.default}; see
    the documentation on that method function}
}
\references{
  Ragin, C. C. 1987. The Comparative Method: Moving Beyond Qualitative and Quantitative Strategies. Berkeley: University of California Press, pp. 118-121.
  
  Schneider, C. Q., Wagemann, C. 2012. Set-Theoretic Methods for the Social Sciences: A Guide to Qualitative Comparative Analysis. Cambridge: Cambridge University Press, chapter 11.3.
}
\author{
  Ioana-Elena Oana
}
\seealso{
  \code{\link[QCA:minimize]{minimize}}
}
\keyword{QCA}
