\name{SAMF}
\alias{SAMF}
\docType{data}
\title{Samford (2010)}
\description{The \code{SAMF} data frame has 61 rows and 4 sets}
\usage{data(SAMF)}
\format{
  A data frame with 61 observations on the following 4 sets.
  \describe{
    \item{\code{Y}}{a numeric vector. Outcome, trade liberalization.}
    \item{\code{G}}{a numeric vector. Condition, lack of weak growth.}
    \item{\code{H}}{a numeric vector. Condition, lack of hyper-inflation}
    \item{\code{HorG}}{a numeric vector. Condition, H or G.}
  }
}
\details{Data are used by Samford (2010) to analyze rapid trade liberalization in Latin America. Data are fuzzy-sets.}
\references{
Samford, S. (2010) "Averting 'Disruption and Reversal': Reassessing the Logic of Rapid Trade Reform in Latin America", Politics and Policy 38(3), pp. 373-407.

Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

}
\examples{data(SAMF)}
\keyword{datasets}
