\name{SC}
\alias{SC}
\docType{data}
\title{Selbst, practicing the truth table algorithm data}

\description{The \code{SC} data frame has 130 rows and 4 sets}

\usage{data(SC)}
\format{
  A data frame with 130 observations on the following 4 sets.
  \describe{
    \item{\code{A}}{a numeric vector. Condition, crisp-set.}
    \item{\code{B}}{a numeric vector. Condition, crisp-set.}
    \item{\code{C}}{a numeric vector. Condition, crisp-set.}
    \item{\code{Y}}{a numeric vector. Condition, crisp-set.}
  }
}
\details{The authors of the data are Carsten Schnieder and Claudius Wagemann. The data are used in the on-line appendix of "Set-Theoretic Methods for the Social Sciences" to practice the truth table algorithm. Data are crisp-sets.}

\references{
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012) How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at www.cambridge.org/schneider-wagemann
}

\examples{data(SC)}
\keyword{datasets}
