\name{SCHF}
\alias{SCHF}
\docType{data}
\title{Schneider et. al. (2010)}
\description{The \code{SCHF} data frame has 76 observations and 7 sets}
\usage{data(SCHF)}
\format{
  A data frame with 76 observations on the following 7 sets.
  \describe{
    \item{\code{EMP}}{a numeric vector. Condition, employment protection.}
    \item{\code{BARGAIN}}{a numeric vector. Condition, collective bargaining.}
    \item{\code{UNI}}{a numeric vector. Condition, university training.}
    \item{\code{OCCUP}}{a numeric vector. Condition, occupational training.}
    \item{\code{STOCK}}{a numeric vector. Condition, stock market size.}
    \item{\code{MA}}{a numeric vector. Condition, mergers and acquisitions.}
    \item{\code{EXPORT}}{a numeric vector. Outcome, export performance in high-tech industries.}
  }
}
\details{Data used by Schneider et. al. (2010) to explain capitalist variety 
and export performance in high-tech industries. Data are fuzzy-sets.}

\references{Schneider, M. R., Schulze-Bentrop, C., Paunescu, M. (2010) "Mapping the institutional capital of high-tech firms: A fuzzy-set analysis of capitalist variety and export performance", Journal of International Business Studies 41, pp. 246-266.
	
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

}
\examples{data(SCHF)}
\keyword{datasets}
