\name{Selbst.disappear}
\alias{Selbst.disappear}
\docType{data}
\title{Selbst, disappearing necessary condition data}
\description{The \code{Selbst.disappear} data frame has 98 rows and 4 variables}
\usage{data(Selbst.disappear)}
\format{
  A data frame with 98 observations on the following 4 variables.
  \describe{
    \item{\code{A}}{a numeric vector. Condition, crisp-set.}
    \item{\code{B}}{a numeric vector. Condition, crisp-set.}
    \item{\code{C}}{a numeric vector. Condition, crisp-set.}
    \item{\code{Y}}{a numeric vector. Outcome, crisp-set.}
  }
}
\details{The authors of the data are Carsten Schnieder and Claudius Wagemann. The data are used in the on-line appendix of "Set-Theoretic Methods for the Social Sciences" to show the disappearance of a necessary condition. Data are crisp-sets.}

\references{
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012) How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at \url{www.cambridge.org/schneider-wagemann}

}
\examples{data(Selbst.disappear)}
\keyword{datasets}
