\name{SetMethods-package}
\alias{SetMethods-package}
\alias{SetMethods}
\docType{package}
\title{SetMethods: A Package Companion to "Set-Theoretic Methods for the Social Sciences"}
\description{SetMethods is a package companion to the book by C. Q. Schneider and C. Wagemann "Set-Theoretic Methods for the Social Sciences", Cambridge University Press. It contains some additional functions not present in other packages and data to replicate the examples in the book and in the Online Appendix.}
\details{
\tabular{ll}{
Package: \tab SetMethods\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2015-06-17\cr
License: \tab GPL-2\cr
}
The package contains some additional functions to perform indirect calibration, calculate parameters of fit and produce XY plots. Furthermore, it contains all the data used in "Set-Theoretic Methods for the Social Sciences". The package is intended to be used to replicate the analyses shown in the Online Appendix of the book.}
\author{
Juraj Medzihorsky, Ioana-Elena Oana, Carsten Q. Schneider, Mario Quaranta

Maintainer: Ioana-Elena Oana <oana_ioana-elena@phd.ceu.edu>
}
\references{
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012). How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at \url{www.cambridge.org/schneider-wagemann}

Schneider, C. Q., Rohlfing, I. 2013. Combining QCA and Process Tracing in Set-Theoretic Multi-Method Research. Sociological Methods Research 42(4): 559-597

}
%\keyword{}
%\seealso{}
%\examples{}
