\name{SetMethodsN-package}
\alias{SetMethodsN-package}
\alias{SetMethodsN}
\docType{package}
\title{
\packageTitle{SetMethodsN}
}
\description{
\packageDescription{SetMethodsN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SetMethodsN}
\packageIndices{SetMethodsN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SetMethodsN}

Maintainer: \packageMaintainer{SetMethodsN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
