\name{STUF}
\alias{STUF}
\title{Fake fuzzy-set student data.}
\description{Fake fuzzy-set football practice data used for Chapter 4 of the Oana et al. (forthcoming) book.}
\usage{data(STUF)}
\format{
  A data frame with 17 observations on the following 3 variables. Conditions for arriving in time for football practice.
  \describe{
    \item{\code{A}}{Condition: Live close by}
    \item{\code{B}}{Condition: Arrive by bike}
    \item{\code{C}}{Condition: Come from punctual familie}
    \item{\code{D}}{Condition: Are seasoned players}
    \item{\code{E}}{Condition: Goalkeepers}
    \item{\code{Y}}{Outcome: Arrive on time for football practice}
  }
}
\references{
Oana, Ioana-Elena, Carsten Q. Schneider, and Eva Thomann (forthcoming). Qualitative Comparative Analysis (QCA) using R: A Gentle Introduction. Cambridge: Cambridge University Press.
}
\examples{data(STUF)}
\keyword{datasets}
