\name{STUR}
\alias{STUR}
\title{Fake raw, student data.}
\description{Fake raw, student data used for Chapter 2 of the Oana et al. (forthcoming) book.}
\usage{data(STUR)}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{\code{MARK}}{Cond.: student mark}
    \item{\code{PARTICIPATION}}{Cond.: quality and correctness of the student's comments in class,}
    \item{\code{PEERS}}{Cond.: extent to which the student engages with and helps peers during interactive in-class teaching exercises}
  }
}
\references{
Oana, Ioana-Elena, Carsten Q. Schneider, and Eva Thomann (forthcoming). Qualitative Comparative Analysis (QCA) using R: A Gentle Introduction. Cambridge: Cambridge University Press.
}
\examples{data(STUR)}
\keyword{datasets}
