\name{theory.fit}
\alias{theory.fit}
\title{
	Function that returns parameters of fit for the intersections between theory and the empirical solution.
}
\description{
		Function that returns parameters of fit for the intersections between theory and the empirical solution.
}
\usage{
theory.fit(theory_eval)
}
\arguments{
	\item{theory_eval}{
		A data frame with the intersections between theory and empirics. 
		Usually the output of the theory.evaluation function should be inputed.
	}
	}
\details{
}
\value{
}
\references{
}
\author{
	Ioana-Elena Oana
}
\note{
}
\seealso{
	\code{\link[QCA:eqmcc]{eqmcc}}
}
\examples{
}
\keyword{QCA}
