\name{THOF}
\alias{THOF}
\docType{data}
\title{Thomann (2015)}
\description{The \code{THOF} data frame has 76 rows and 7 sets. The study analyses how European Union (EU) member states adapt EU directives
to domestic contexts during transposition and asks how and why fully compliant countries 'customize' EU directives.}
\usage{data(THOF)}
\format{
  A data frame with 76 observations on the following 7 sets.
  \describe{
    \item{\code{RESP}}{Cond.: Responsive regulatory mode}
    \item{\code{SAL}}{Cond.: Salient issue }
    \item{\code{RES}}{Cond.: Domestic resistance}
    \item{\code{VPO}}{Cond.: Many veto points }
    \item{\code{VPL}}{Cond.: Many veto players }
    \item{\code{COERC}}{Cond.: Coercive interventionist style}
    \item{\code{CUSTOM}}{Outcome: Extensive customization}
  }
}
\details{Data used by Thomann (2015) to explain customization. Data are fuzzy-sets.}

\references{
Thomann, E. 2015. Customizing Europe: Transposition as bottom-up implementation. Journal of European Public Policy 22(10): 1368-1387.
}
\examples{data(THOF)}
\keyword{datasets}
