\name{VISC}
\alias{VISC}
\docType{data}
\title{Vis (2009), crisp set data}
\description{The \code{VISC} data frame has 25 rows and 4 sets}
\usage{data(VISC)}
\format{
  A data frame with 25 observations on the following 4 sets.
  \describe{
    \item{\code{P}}{a numeric vector. Condition, weak political positions, with parties in government expecting losses at the next election.}
    \item{\code{S}}{a numeric vector. Condition, deteriorating economic situation.}
    \item{\code{R}}{a numeric vector. Condition, government dominated by parties from the right of the political spectrum.}
    \item{\code{U}}{a numeric vector. Outcome, unpopular reform.}
  }
}
\details{Data are used by Vis (2009) to analyze the pursuit of unpopular reforms by governments. Data are crisp-sets.}
\references{
Vis, B. (2009) "Government and Unpopular Social Policy Reforms: Biting the Bullet or Steering Clear?", European Journal of Political Research 48, pp. 31-57.

Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge.

}
\examples{data(VISC)}
\keyword{datasets}
