\name{cases.suf.typ.most}
\alias{cases.suf.typ.most}
\title{
	Lists most typical cases with regards to sufficiency.
}
\description{
	Function extracts most typical cases with regards to sufficiency from an object of class "qca".
}
\usage{
cases.suf.typ.most(results, outcome, neg.out=TRUE, intermed=TRUE, sol = 1)
}
\arguments{
	\item{results}{
		An object of class "qca".
	}
	\item{outcome}{
		A character string with the name of the outcome.
	}
	\item{neg.out}{
		Logical. Should the negated outcome be used?
	}
	\item{intermed}{
		Logical. Should the intermediate solution be used?
	}
	\item{sol}{
		A numeric vector where the first number indicates the number of the
		solution according to the order in the "qca" object.
	}
}
\details{
}
\value{
}
\references{
}
\author{
	Juraj Medzihorsky and Ioana-Elena Oana
}
\note{
}
\seealso{
	\code{\link[QCA:eqmcc]{eqmcc}}
}
\examples{}
\keyword{QCA}
