\name{matches.suf.typdcn}
\alias{matches.suf.typdcn}
\title{
	Match typical cases with cases deviant with regards to necessity.
}
\description{
	Matches typical cases with cases deviant with regards to necessity.
}
\usage{
matches.suf.typdcn(results, outcome, solution = 1, term = NULL, max_pairs = 5)
}
\arguments{
	\item{results}{
		An object of class "qca".
	}
	\item{outcome}{
		A character string with the name of the outcome.
	}
	\item{solution}{
		A numeric vector where the first number indicates the number of the
		solution according to the order in the "qca" object.
	}
	\item{term}{
		Which term?
	}
	\item{max_pairs}{
		Maximum number of pairs to extract.
	}
}
\details{
}
\value{
}
\references{
}
\author{
	Juraj Medzihorsky
}
\note{
}
\seealso{
	\code{\link[QCA:eqmcc]{eqmcc}}
}
\examples{}
\keyword{QCA}
