\name{SetMethods-package}
\alias{SetMethods-package}
\alias{SetMethods}
\docType{package}
\title{Functions for Set-Theoretic Multi-Method Research and Advanced QCA}
\description{This initiated as a package companion to the book by C. Q. Schneider and C. Wagemann "Set-Theoretic Methods for the Social Sciences", Cambridge University Press. It has now grown to include functions for performing set-theoretic multi-method research (Schneider and Rohlfing 2013), QCA for clustered data (Garcia-Castro and Arino 2013), theory evaluation, and Enhanced Standard Analysis. Additionally it includes data to replicate the examples in the book and in the online appendix.}
\details{
\tabular{ll}{
Package: \tab SetMethods\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2016-08-09\cr
License: \tab GPL-2\cr
}
The package contains functions to perform set-theoretic multi-method research, theory evaluation, QCA for clustered data, Enhanced Standard Analyses, indirect calibration, calculate parameters of fit and produce XY plots. Furthermore, it contains all the data used in "Set-Theoretic Methods for the Social Sciences" and more.}
\author{
Juraj Medzihorsky [aut], Ioana-Elena Oana [aut, cre], Carsten Q. Schneider [aut], Mario Quaranta [aut]

Maintainer: Ioana-Elena Oana <oana_ioana-elena@phd.ceu.edu>
}
\references{
Schneider, C. Q., Wagemann, C. (2012) Set-Theoretic Methods for the Social Sciences, Cambridge University Press: Cambridge

Schneider, C. Q., Wagemann, C., Quaranta, M. (2012). How To... Use Software for Set-Theoretic Analysis. Online Appendix to "Set-Theoretic Methods for the Social Sciences". Available at \url{www.cambridge.org/schneider-wagemann}

Schneider, C. Q., Rohlfing, I. 2013. Combining QCA and Process Tracing in Set-Theoretic Multi-Method Research. Sociological Methods Research 42(4): 559-597

Garcia-Castro, A., Arino, M. A.. 2013. A General Approach to Panel Data Set-Theoretic Research. COMPASSS Working Paper 2013-76

}
%\keyword{}
%\seealso{}
%\examples{}
